refactor(message): centralize message reference models
