+++ src-tauri/src/commands/actions.rs
@@ -3,8 +3,8 @@ use tauri::{Emitter, State};
 use crate::db;
 use crate::error::{Error, Result};
 use crate::event::tauri::{emit_folders_changed, emit_messages_changed};
-use crate::mail::compat::{BackendMessageRef, BodyLocation};
 use crate::mail::imap::{ImapConfig, ImapConnection};
+use crate::message::{BackendMessageRef, BodyLocation};
 use crate::ops::flags::{AccountReadScope, FlagMutation, FlagTarget};
 use crate::ops::queue::{MailOp, OpEntry, OpPriority};
 use crate::state::AppState;
