+++ src-tauri/src/commands/mail.rs
@@ -5,10 +5,10 @@ use crate::db;
 use crate::db::messages::{MessageSummary, ThreadedPage};
 use crate::error::{Error, Result};
 use crate::event::tauri::{emit_folders_changed, emit_messages_changed};
-use crate::mail::compat::{BackendMessageRef, BodyLocation};
 use crate::mail::imap::ImapConfig;
 use crate::mail::parser;
 use crate::mail::search::{SearchHit, SearchQuery};
+use crate::message::{BackendMessageRef, BodyLocation};
 use crate::state::AppState;
 
 /// Check if an IP address is in a private/reserved range (SSRF protection).
