+++ src-tauri/src/mail/mod.rs
@@ -1,7 +1,6 @@
 pub mod autoconfig;
 pub mod caldav;
 pub mod carddav;
-pub mod compat;
 pub mod dav_http;
 pub mod google;
 pub mod graph;
